+++ Modules~/HapticModule/Scripts/Runtime/Models/HapticModel.cs
@@ -1,0 +1,9 @@
+namespace Modules.HapticModule.Models
+{
+    public class HapticModel : IHapticModel
+    {
+        public bool IsEnabled { get; private set; } = true;
+
+        public void SetEnabled(bool on) => IsEnabled = on;
+    }
+}