+++ Modules~/HapticModule/Scripts/Runtime/RootsContexts/HapticServiceRoot.cs
@@ -1,0 +1,9 @@
+using FlowIoC.BaseModule.Root;
+
+namespace Modules.HapticModule.RootsContexts
+{
+    public class HapticServiceRoot : Root<HapticServiceContext>
+    {
+        
+    }
+}