+++ Modules~/HapticModule/zTestModules/HapticTestModule/Scripts/Runtime/RootsContexts/HapticTestRoot.cs
@@ -1,0 +1,11 @@
+#if UNITY_EDITOR
+using FlowIoC.BaseModule.Root;
+
+namespace Modules.HapticModule.HapticTestModule.RootsContexts
+{
+    public class HapticTestRoot : Root<HapticTestContext>
+    {
+        
+    }
+}
+#endif